+++ abuf.h
@@ -3,8 +3,8 @@
 /* append buffer */
 
 typedef struct abuf {
-  size_t len;
-  size_t size;
+  unsigned long long len;
+  unsigned long long size;
   char *buf;
 } abuf;
 
@@ -12,8 +12,8 @@ void ab_init(struct abuf *ab);
 
 int ab_charcat(struct abuf *ab, char c);
 
-int ab_strcat(struct abuf *ab, const char *s, size_t len);
+int ab_strcat(struct abuf *ab, const char *s, unsigned long long len);
 
-void ab_pop(struct abuf *ab, ssize_t times);
+void ab_pop(struct abuf *ab, unsigned long long times);
 
 void ab_free(struct abuf *ab);
